+++ P2/include/instance.hpp
@@ -16,9 +16,9 @@
 
 class IdHelper {
 private:
-    unsigned _current_id = 0;
+    unsigned long long _current_id = 0;
 public:
-    Result get_id(unsigned &ret);
+    Result get_id(unsigned long long &ret);
 };
 
 /* Instance is a class that, given a nonnegative integer n, and a nx2 array 
@@ -30,24 +30,24 @@ public:
 class Instance {
 private:
     std::vector<Rectangle> _rectangles;     // Rectangles that make up the instance.
-    unsigned _n_rectangles;                   // Number of rectangles in instance.
+    unsigned long long _n_rectangles;                   // Number of rectangles in instance.
     bool _rectangles_set = false;           // Have the rectangles been set yet?
 
     IdHelper _id_gen;                       // Object that generates IDs.
 
     /* Add a new Rectangle of given width and height to the instance. */
-    Result add_rectangle(unsigned int width, unsigned int height);
+    Result add_rectangle(unsigned long long width, unsigned long long height);
 
 public:
     /* Set Rectangles for this instances given an nx2 array: nx(width height). */
-    Result set_rectangles(unsigned n_rectangles, unsigned int **rectangle_dims);
+    Result set_rectangles(unsigned long long n_rectangles, unsigned long long **rectangle_dims);
 
     /* Verify the instance. */
     Result verify() const;
 
     /* Accessors. */
     const std::vector<Rectangle> &rectangles() const;
-    unsigned n_rectangles() const;
+    unsigned long long n_rectangles() const;
 
     /* IO-functions for testing. */
     Result print_rectangles() const;
