+++ P2/src/instance.cpp
@@ -13,15 +13,15 @@
 
 using namespace std;
 
-Result Instance::set_rectangles(unsigned n_rectangles, 
-                                unsigned **rectangle_dims)
+Result Instance::set_rectangles(unsigned long long n_rectangles, 
+                                unsigned long long **rectangle_dims)
 {
     if (_rectangles_set)
         return FAIL;
 
     _n_rectangles = n_rectangles;
     Result res;
-    for (unsigned i = 0; i < n_rectangles; i++) {
+    for (unsigned long long i = 0; i < n_rectangles; i++) {
         res = add_rectangle(rectangle_dims[i][0], rectangle_dims[i][1]);
         if (res == FAIL)
             return FAIL;
@@ -31,9 +31,9 @@ Result Instance::set_rectangles(unsigned n_rectangles,
     return SUCCESS;
 }
 
-Result Instance::add_rectangle(unsigned width, unsigned height) {
+Result Instance::add_rectangle(unsigned long long width, unsigned long long height) {
     /* First get an id for the rectangle if possible. */
-    unsigned id;
+    unsigned long long id;
     if (_id_gen.get_id(id) == FAIL)
         return FAIL;
 
@@ -51,7 +51,7 @@ Result Instance::verify() const {
     /* Check that the dimensions of each rectangle are in bounds,
      * and that the id of each rectangle corresponds to their location
      * in the array, thus automatically checking that ids are unique. */
-    for (unsigned i = 0; i < _n_rectangles; i++) {
+    for (unsigned long long i = 0; i < _n_rectangles; i++) {
         if (_rectangles[i].height() > MAX_RECT_DIM ||
             _rectangles[i].width() > MAX_RECT_DIM  ||
             _rectangles[i].id() != i)
@@ -66,7 +66,7 @@ const vector<Rectangle> &Instance::rectangles() const {
     return _rectangles;
 }
 
-unsigned Instance::n_rectangles() const { return _n_rectangles; }
+unsigned long long Instance::n_rectangles() const { return _n_rectangles; }
 
 Result Instance::print_rectangles() const {
     if (!_rectangles_set) {
@@ -74,13 +74,13 @@ Result Instance::print_rectangles() const {
         exit(1);
     }
 
-    for (unsigned i = 0; i < _n_rectangles; i++) {
+    for (unsigned long long i = 0; i < _n_rectangles; i++) {
         _rectangles[i].print_rectangle();
     }
     return SUCCESS;
 }
 
-Result IdHelper::get_id(unsigned &ret) { 
+Result IdHelper::get_id(unsigned long long &ret) { 
     if (_current_id == SIZE_MAX)
         return FAIL;
