+++ P2/src/util.cpp
@@ -13,38 +13,38 @@
 
 using namespace std;
 
-Result Free2DArray(unsigned** array, unsigned n) {
-    for (unsigned i = 0; i < n; i++)
+Result Free2DArray(unsigned long long** array, unsigned long long n) {
+    for (unsigned long long i = 0; i < n; i++)
             free(array[i]);
     free(array);
     return SUCCESS;
 }
 
 /* TODO: add exception. */
-unsigned Fact(unsigned n) {
-    unsigned ret = 1;
-    for(unsigned i = 1; i <= n; i++)
+unsigned long long Fact(unsigned long long n) {
+    unsigned long long ret = 1;
+    for(unsigned long long i = 1; i <= n; i++)
         ret *= i;
     return ret;
 }
 
-void print_vector(vector<unsigned> v) {
+void print_vector(vector<unsigned long long> v) {
     for (auto const& c : v)
         cout << c << ' ';
     cout << endl;
 }
 
 /* TODO: add exception. */
-unsigned subtract(unsigned a, unsigned b) {
+unsigned long long subtract(unsigned long long a, unsigned long long b) {
     if (a < b)
         cout << "Util: Warning subtract lead to underflow" << endl;
     return a - b;
 }
 
-bool between_right(unsigned a, unsigned b, unsigned c) {
+bool between_right(unsigned long long a, unsigned long long b, unsigned long long c) {
     return ((a <= b) && (b < c));
 }
 
-bool between_left(unsigned a, unsigned b, unsigned c) {
+bool between_left(unsigned long long a, unsigned long long b, unsigned long long c) {
     return ((a < b) && (b <= c));
 }
